"bugfix Zip.file"

+++ commands/zip9.py
@@ -3,7 +3,7 @@
 """Internal module to work with zip archives
 """
 # https://code.tutsplus.com/ru/tutorials/compressing-and-extracting-files-in-python--cms-26816
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
 
 class Zip:
@@ -16,13 +16,15 @@ class Zip:
         if not File.exist(input_file):
             raise IOError(input_file + " is not file")
 
-        temp_file = output_zip+".tmp"
-        File.delete(temp_file, quiet=True)
+        if not File.exist(output_zip) and "w" not in mode:
+            temp_file = output_zip+".tmp"
+            File.delete(temp_file, quiet=True)
 
         with zipfile.ZipFile(temp_file, mode=mode) as zip:
             zip.write(input_file, arcname=arcname, compress_type=zipfile.ZIP_DEFLATED)
 
-        File.move(temp_file, output_zip)
+        if not File.exist(output_zip) and "w" not in mode:
+            File.move(temp_file, output_zip)
 
     @staticmethod
     def dir(input_dir, output_zip, quiet=True):
